merge_wordsの場所を移動 fix #38
