+++ include/executor.h
@@ -3,17 +3,12 @@
 # define EXECUTOR_H
 
 # include "ast.h"
+# include "expander.h"
 # include "minishell.h"
 # include "token.h"
 
 void	run_cmd(t_cmd *cmd, t_mgr *mgr);
 void	exec_cmd(t_cmd *cmd, t_mgr *mgr);
-
-/* exec_pipe.c */
 void	exec_pipe(t_cmd *cmd, t_mgr *mgr);
-/* merge_words.c */
-void	merge_words(t_word *word_list);
-/* expand_exit_status.c */
-void	expand_word_list_for_exit_status(t_word *word_list, int exit_status);
 
 #endif
