+++ include/expander.h
@@ -12,6 +12,7 @@ void	expand_env(char **word, char **cur_ptr, t_hash_table *env_table);
 void	run_expansion(t_cmd *cmd, t_hash_table *env_table);
 // Util
 char	**ft_split(char const *s, char c);
-/* exp_split_word.c */
 void	split_word_token(t_word *word);
+void	merge_words(t_word *word_list);
+void	expand_word_list_for_exit_status(t_word *word_list, int exit_status);
 #endif
